# Conflicts:
#	build.xml
#	src/haven/Buff.java
#	src/haven/Defer.java
#	src/haven/Fightsess.java
#	src/haven/GItem.java
#	src/haven/GameUI.java
#	src/haven/Glob.java
#	src/haven/Gob.java
#	src/haven/GobHealth.java
#	src/haven/HavenPanel.java
#	src/haven/KinInfo.java
#	src/haven/Listbox.java
#	src/haven/MainFrame.java
#	src/haven/MapView.java
#	src/haven/MenuGrid.java
#	src/haven/OCache.java
#	src/haven/OptWnd.java
#	src/haven/RenderList.java
#	src/haven/ResDrawable.java
#	src/haven/RootWidget.java
#	src/haven/Session.java
#	src/haven/UI.java
#	src/haven/Widget.java

+++ src/haven/Buff.java
@@ -118,7 +118,7 @@ public class Buff extends Widget implements ItemInfo.ResOwner, Bufflist.Managed 
 
     public void draw(GOut g) {
 	g.chcolor(255, 255, 255, a);
-	Double ameter = (this.ameter >= 0) ? (this.ameter / 100.0) : ameteri.get();
+	Double ameter = (this.ameter >= 0) ? Double.valueOf(this.ameter / 100.0) : ameteri.get();
 	int ameteri = 0;
 	if(ameter != null) {
 	    ameteri = (int) (100*ameter);
